api/requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
sqlmodel
python-decouple